revcopy-backend-main/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# HTTP Client & Web Scraping
aiohttp==3.9.1
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# AI & ML
openai==1.3.7
anthropic==0.7.7

# Rate Limiting
slowapi==0.1.9

# Caching & Background Tasks
redis==5.0.1
celery==5.3.4

# Monitoring & Logging
structlog==23.2.0

# Environment & Config
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2

# Data Validation
email-validator==2.1.0

# File handling
aiofiles==23.2.1

# Additional utilities
click==8.1.7

# Security
bcrypt==4.1.2

# Async utilities
aioredis==2.0.1

# Time handling
python-dateutil==2.8.2
locust==2.17.0

# Database optimization
sqlparse==0.4.4
pg8000==1.30.3

# Geo-location
geoip2==4.7.0

# Compression
lz4==4.3.2
zstandard==0.22.0

# Serialization
msgpack==1.0.7
orjson==3.9.10

# Networking
websockets==12.0
sse-starlette==1.6.5

# Advanced security
argon2-cffi==23.1.0
scrypt==0.8.20

# Database connection pooling
sqlalchemy-pool==0.7.0

# Advanced HTTP features
httptools==0.6.1
h11==0.14.0

# Development tools
ipython==8.17.2
jupyter==1.0.0

# Memory optimization
pymongo==4.6.0
motor==3.3.2

# Advanced caching
django-redis==5.4.0
redis-py-cluster==2.1.3

# Task scheduling
apscheduler==3.10.4
python-crontab==3.0.0

# API documentation
fastapi-utils==0.2.1
openapi-spec-validator==0.7.1

# Advanced logging
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0

# Monitoring and alerting
alertmanager-webhook-logger==0.4.0

# Content processing
python-magic==0.4.27
chardet==5.2.0

# Configuration validation
cerberus==1.3.4
voluptuous==0.13.1

# Advanced testing
factory-boy==3.3.0
faker==20.1.0
responses==0.24.1

# Documentation generation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Code analysis
radon==6.0.1
xenon==0.9.1

# Performance profiling
line-profiler==4.1.1
scalene==1.5.26

# Database introspection
sqlalchemy-migrate==0.13.0
sqlalchemy-utils==0.41.1

# Advanced HTTP client
httpx-oauth==0.10.2
authlib==1.2.1

# Message queuing
pika==1.3.2
kombu==5.3.4

# Advanced serialization
pydantic-extra-types==2.2.0
pydantic-core==2.14.6

# Content type detection
python-magic-bin==0.4.14

# Advanced datetime handling
pytz==2023.3.post1
tzdata==2023.3

# Enhanced error handling
sentry-sdk[fastapi]==1.38.0
rollbar==0.16.3

# Advanced file processing
python-docx==1.1.0
python-pptx==0.6.23

# Network utilities
dnspython==2.4.2
netaddr==0.10.1

# Advanced validation
wtforms==3.1.1
wtforms-validators==1.1.2

# Database migrations
alembic-utils==0.8.2
alembic-postgresql-enum==1.1.2

# Advanced monitoring
statsd-telegraf==0.4.0
py-cpuinfo==9.0.0

# Content delivery
boto3-stubs==1.34.0
botocore-stubs==1.34.0

# Advanced security scanning
semgrep==1.48.0
bandit-sarif-formatter==1.1.1

# Performance benchmarking
pytest-benchmark==4.0.0
py-spy==0.3.14

# Advanced compression
brotli==1.1.0
zlib-ng==0.3.0

# Database performance
sqlalchemy-continuum==1.4.0
sqlalchemy-searchable==1.5.0

# Advanced logging formatters
colorlog==6.8.0
termcolor==2.3.0

# Content analysis
textstat==0.7.3
readability==0.3.1

# Advanced networking
yarl==1.9.3
multidict==6.0.4

# Background task monitoring
celery-progress==0.2.0
celery-beat-scheduler==2.5.0

# Advanced file handling
pathlib2==2.3.7
scandir==1.10.0

# Performance optimization
cython==3.0.6
numba==0.58.1

# Advanced data structures
sortedcontainers==2.4.0
bidict==0.22.1

# Content processing
ftfy==6.1.1
unidecode==1.3.7

# Advanced HTTP handling
urllib3==2.1.0
certifi==2023.11.17

# Database connection management
pgbouncer==1.21.0
connection-pool==0.0.3

# Advanced caching strategies
dogpile.cache==1.3.0
beaker==1.12.1

# Enhanced security
oauthlib==3.2.2
cryptography==41.0.7

# Advanced content generation
markovify==0.9.4
gensim==4.3.2

# Performance monitoring
py-cpuinfo==9.0.0
psutil==5.9.6

# Advanced configuration
hydra-core==1.3.2
omegaconf==2.3.0

# Content validation
jsonschema==4.20.0
schema==0.7.5

# Advanced testing frameworks
hypothesis==6.92.1
pytest-xdist==3.5.0

# Database optimization
sqlalchemy-mixins==2.0.5
sqlalchemy-serializer==1.4.1

# Advanced monitoring
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-prometheus==1.12.0rc1

# Content processing
python-multipart==0.0.6
email-validator==2.1.0

# Advanced security
python-jose[cryptography]==3.3.0
itsdangerous==2.1.2 